refactor: 데이터 저장 없이 촬영 사진 바로 임시 파일 저장하도록 수정

+++ mirroringBooth/mirroringBooth/Device/Camera/Streaming/CameraManageable.swift
@@ -11,7 +11,7 @@ protocol CameraManageable: AnyObject {
     var rawData: ((CMSampleBuffer) -> Void)? { get set }
     var onEncodedData: ((Data) -> Void)? { get set }
     var onTransferCompleted: (() -> Void)? { get set }
-    var onAllPhotosStored: ((Int) -> Void)? { get set }
+    var onAllPhotosStored: (() -> Void)? { get set }
 
     func startSession()
     func stopSession()
